refactor(web): #35 - move Argument component folder inside ResourceNode component folder
